runners/genlayer-py-std/pyproject.toml
[tool.poetry]
name = "genvm-sdk-python"
version = "v0.0.1"
description = ""
authors = ["Kira Prokopenko <kira@yeager.ai>", "Yeager AI Corp"]
#readme = "README.md"
packages = [
    { include = "genlayer", from = "src" },
    { include = "genlayer_embeddings", from = "src-emb" },
    { include = "onnx", from = "src-emb" },
]

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-xdist = "^3.6.1"
pytest-cov = "^6.0.0"
sentence-transformers = "^3.2.1"
eth-abi = "^5.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
addopts = "--cov=src/genlayer/py --cov-report term-missing --cov-fail-under=75"
